Change everything from Long to double, so you can place shifts not on the grid

+++ optashift-employee-rostering-gwtui/src/main/java/org/optaplanner/openshift/employeerostering/gwtui/client/pages/rotation/ShiftTemplateBlobView.java
@@ -104,7 +104,7 @@ public class ShiftTemplateBlobView implements BlobView<LocalDateTime, ShiftTempl
         label.textContent = (blob.getShiftTemplateView().getRotationEmployeeId() != null) ? blob.getRotationEmployee().getName() : "Unassigned";
     }
 
-    private void onResize(final Long newSizeInGridPixels,
+    private void onResize(final double newSizeInGridPixels,
                           final CollisionState dragState) {
 
         refresh();
@@ -115,7 +115,7 @@ public class ShiftTemplateBlobView implements BlobView<LocalDateTime, ShiftTempl
         }
     }
 
-    private void onDrag(final Long newPositionInGridPixels,
+    private void onDrag(final double newPositionInGridPixels,
                         final CollisionState dragState) {
 
         refresh();
