+++ optashift-employee-rostering-gwtui/src/main/java/org/optaplanner/openshift/employeerostering/gwtui/client/pages/spotroster/ShiftBlob.java
@@ -49,7 +49,7 @@ public class ShiftBlob extends TimeslotBlob {
     }
 
     @Override
-    public void setSizeInGridPixels(final long sizeInGridPixels) {
+    public void setSizeInGridPixels(final double sizeInGridPixels) {
         shift.setEndDateTime(getScale().toScaleUnits(getScale()
                 .toGridPixels(getPositionInScaleUnits()) + sizeInGridPixels));
     }
